app/pyproject.toml
[project]
name = "dr-ai-palletizer"
version = "0.1.0"
description = "Main application integrating simulation, motion planning, and AI"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.29.0",
    "Pillow>=11.0.0",
    "httpx>=0.27",
    "openai>=1.40",
    "pydantic-settings>=2.0",
    "numpy>=1.26",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dr_ai_palletizer"]
